Eliminate (trivial) warning on MinGW
git-svn-id: http://svn.mythtv.org/svn/trunk@15546 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/compat.h
@@ -205,7 +205,7 @@ inline const char *dlerror(void)
                         MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
                         errStr, DLERR_MAX - 1, NULL))
         snprintf(errStr, DLERR_MAX - 1,
-                 "dlopen()/dlsym() caused error %d", errCode);
+                 "dlopen()/dlsym() caused error %d", (int)errCode);
 
     return errStr;
 }
